requirements.txt
aiofiles==24.1.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
click==8.1.7
cryptography==45.0.4
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.10.0
pydantic_core==2.27.2
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.12.2
uvicorn==0.34.0
